+++ src/combinator/provided/bytes/wrap.rs
@@ -1,6 +1,6 @@
 use crate::{
   action::{Action, Input, Output},
-  combinator::{create_closure_combinator, Combinator},
+  combinator::{create_closure_combinator, Combinator, Contextual},
   digest::Digest,
   instant::Instant,
 };
@@ -13,10 +13,16 @@ create_closure_combinator!(Wrap, "See [`wrap`] and [`wrap`].");
 
 macro_rules! impl_wrap {
   ($name:ident, $assert:ident, $text:ty) => {
-    unsafe impl<Value, F: Fn(&Instant<&$text>) -> Option<Output<Value>>> Action for $name<F> {
+    unsafe impl<
+        State,
+        Heap,
+        Value,
+        F: Fn(Input<&Instant<&$text>, &mut State, &mut Heap>) -> Option<Output<Value>>,
+      > Action for Contextual<$name<F>, State, Heap>
+    {
       type Text = $text;
-      type State = ();
-      type Heap = ();
+      type State = State;
+      type Heap = Heap;
       type Value = Value;
 
       #[inline]
@@ -24,11 +30,11 @@ macro_rules! impl_wrap {
         &self,
         input: Input<&Instant<&Self::Text>, &mut Self::State, &mut Self::Heap>,
       ) -> Option<Output<Self::Value>> {
-        let output = (self.inner)(input.instant);
-        $assert!(output.as_ref().map_or(true, |output| input
-          .instant
-          .rest()
-          .validate(output.digested)));
+        let instant = input.instant;
+        let output = (self.inner.inner)(input);
+        $assert!(output
+          .as_ref()
+          .map_or(true, |output| instant.rest().validate(output.digested)));
         output
       }
     }
@@ -37,6 +43,8 @@ macro_rules! impl_wrap {
 
 impl_wrap!(WrapUnchecked, debug_assert, [u8]);
 impl_wrap!(Wrap, assert, [u8]);
+
+// TODO: for non-contextual version, can we just return Combinator<Wrap>?
 
 /// Wrap a closure or function to create a [`Combinator`] for bytes.
 ///
@@ -51,14 +59,17 @@ impl_wrap!(Wrap, assert, [u8]);
 /// # use whitehole::action::{Output, Action};
 /// # fn t() -> Combinator<impl Action<[u8]>> {
 /// // eat the next byte if it exists
-/// unsafe { wrap_unchecked(|instant| instant.accept(1)) }
+/// unsafe { wrap_unchecked(|input| input.instant.accept(1)) }
 /// # }
 /// ```
 #[inline]
-pub const unsafe fn wrap_unchecked<Value, F: Fn(&Instant<&[u8]>) -> Option<Output<Value>>>(
+pub const unsafe fn wrap_unchecked<
+  Value,
+  F: Fn(Input<&Instant<&[u8]>, &mut (), &mut ()>) -> Option<Output<Value>>,
+>(
   f: F,
-) -> Combinator<WrapUnchecked<F>> {
-  Combinator::new(WrapUnchecked::new(f))
+) -> Combinator<Contextual<WrapUnchecked<F>, (), ()>> {
+  Combinator::new(Contextual::new(WrapUnchecked::new(f)))
 }
 
 // TODO: merge dup code
@@ -75,14 +86,17 @@ pub const unsafe fn wrap_unchecked<Value, F: Fn(&Instant<&[u8]>) -> Option<Outpu
 /// # use whitehole::action::{Output, Action};
 /// # fn t() -> Combinator<impl Action<[u8]>> {
 /// // eat the next byte if it exists
-/// wrap(|instant| instant.accept(1))
+/// wrap(|input| input.instant.accept(1))
 /// # }
 /// ```
 #[inline]
-pub const fn wrap<Value, F: Fn(&Instant<&[u8]>) -> Option<Output<Value>>>(
+pub const fn wrap<
+  Value,
+  F: Fn(Input<&Instant<&[u8]>, &mut (), &mut ()>) -> Option<Output<Value>>,
+>(
   f: F,
-) -> Combinator<Wrap<F>> {
-  Combinator::new(Wrap::new(f))
+) -> Combinator<Contextual<Wrap<F>, (), ()>> {
+  Combinator::new(Contextual::new(Wrap::new(f)))
 }
 
 #[cfg(test)]
@@ -113,7 +127,7 @@ mod tests {
 
   #[test]
   fn combinator_bytes_wrap_unchecked() {
-    let c = unsafe { wrap_unchecked(|instant| instant.accept(1)) };
+    let c = unsafe { wrap_unchecked(|input| input.instant.accept(1)) };
     helper(c, b"1", 1);
 
     // ensure the combinator is copyable and clone-able
@@ -121,14 +135,17 @@ mod tests {
     let _ = c.clone();
 
     // ensure the combinator is debuggable
-    assert_eq!(format!("{:?}", c), "Combinator { action: WrapUnchecked }");
+    assert_eq!(
+      format!("{:?}", c),
+      "Combinator { action: Contextual(WrapUnchecked) }"
+    );
   }
 
   #[test]
   #[should_panic]
   fn combinator_bytes_wrap_unchecked_overflow() {
     helper(
-      unsafe { wrap_unchecked(|instant| instant.accept_unchecked(4).into()) },
+      unsafe { wrap_unchecked(|input| input.instant.accept_unchecked(4).into()) },
       b"1",
       0,
     );
@@ -136,7 +153,7 @@ mod tests {
 
   #[test]
   fn combinator_bytes_wrap() {
-    let c = wrap(|instant| instant.accept(1));
+    let c = wrap(|input| input.instant.accept(1));
     helper(c, b"1", 1);
 
     // ensure the combinator is copyable and clone-able
@@ -144,14 +161,17 @@ mod tests {
     let _ = c.clone();
 
     // ensure the combinator is debuggable
-    assert_eq!(format!("{:?}", c), "Combinator { action: Wrap }");
+    assert_eq!(
+      format!("{:?}", c),
+      "Combinator { action: Contextual(Wrap) }"
+    );
   }
 
   #[test]
   #[should_panic]
   fn combinator_bytes_wrap_overflow() {
     helper(
-      wrap(|instant| unsafe { instant.accept_unchecked(4) }.into()),
+      wrap(|input| unsafe { input.instant.accept_unchecked(4) }.into()),
       b"1",
       0,
     );